+++ sceptre/resolvers/project_variables.py
@@ -2,6 +2,9 @@
 
 import os
 import yaml
+from warnings import warn, simplefilter
+
+from colorama import Fore, Style
 
 from sceptre.resolvers import Resolver
 
@@ -24,6 +27,16 @@ class ProjectVariables(Resolver):
         :returns: Value at argument location.
         :rtype: str
         """
+        simplefilter("always", DeprecationWarning)
+        warn(
+            "{0}The project_variables resolver has been deprecated, and will"
+            "be removed in a later version of Sceptre. Depending on your "
+            "use case, you may find user variables appropiate. "
+            "Example: sceptre --var ""iam_role=<your iam role>"" <COMMAND>{1}"
+            .format(Fore.YELLOW, Style.RESET_ALL),
+            DeprecationWarning,
+            stacklevel=2
+        )
         file_path = os.path.join(
             self.environment_config.sceptre_dir,
             self.argument