+++ tests/test_cli.py
@@ -43,7 +43,7 @@ class TestCli(object):
         mock_get_env.assert_called_with(sentinel.cwd, "dev", {})
 
         assert result.output == "{0}\n".format(
-            mock_get_env.return_value.stacks["vpc"].template.cfn
+            mock_get_env.return_value.stacks["vpc"].template.body
         )
 
     @patch("sceptre.cli.os.getcwd")
@@ -272,7 +272,8 @@ class TestCli(object):
                         "Replacement": "True",
                         "PhysicalResourceId": "igw-04a59561",
                         "Action": "Remove",
-                        "LogicalResourceId": "InternetGateway"
+                        "LogicalResourceId": "InternetGateway",
+                        "Scope": []
                     }
                 }
             ],