+++ tests/test_template.py
@@ -8,6 +8,7 @@ import threading
 import pytest
 from mock import patch, sentinel, Mock
 
+from freezegun import freeze_time
 from botocore.exceptions import ClientError
 
 from sceptre.template import Template
@@ -36,7 +37,7 @@ class TestTemplate(object):
         assert self.template.path == "/folder/template.py"
         assert self.template.name == "template"
         assert self.template.sceptre_user_data == {}
-        assert self.template._cfn is None
+        assert self.template._body is None
 
     def test_repr(self):
         representation = self.template.__repr__()
@@ -44,25 +45,16 @@ class TestTemplate(object):
             "name='template', path='/folder/template.py'"\
             ", sceptre_user_data={})"
 
-    def test_cfn_with_cache(self):
-        self.template._cfn = sentinel.cfn
-        cfn = self.template.cfn
-        assert cfn == sentinel.cfn
+    def test_body_with_cache(self):
+        self.template._body = sentinel.body
+        body = self.template.body
+        assert body == sentinel.body
 
-    @patch("sceptre.template.Template._get_cfn")
-    def test_cfn_without_cache(self, mock_get_cfn):
-        self.template._cfn = None
-        mock_get_cfn.return_value = sentinel.cfn
-        cfn = self.template.cfn
-        assert cfn == sentinel.cfn
-
-    @patch("sceptre.template._get_time_stamp")
-    @patch("sceptre.template.Template._create_bucket")
-    def test_upload_to_s3_with_valid_arguments(
-            self, mock_create_bucket, mock_get_timestamp
-    ):
-        self.template._cfn = '{"template": "mock"}'
-        mock_get_timestamp.return_value = "2016-10-10-14-32-30-0-Z"
+    @freeze_time("2012-01-01")
+    @patch("sceptre.template.Template._bucket_exists")
+    def test_upload_to_s3_with_valid_arguments(self, mock_bucket_exists):
+        self.template._body = '{"template": "mock"}'
+        mock_bucket_exists.return_value = True
 
         url = self.template.upload_to_s3(
             region="eu-west-1",
@@ -75,12 +67,9 @@ class TestTemplate(object):
 
         expected_template_key = (
             "prefix/eu-west-1/environment/path/"
-            "stack-name-2016-10-10-14-32-30-0-Z.json"
+            "stack-name-2012-01-01-00-00-00-000000Z.json"
         )
 
-        mock_create_bucket.assert_called_once_with(
-            "eu-west-1", "bucket-name", self.connection_manager
-        )
         self.connection_manager.call.assert_called_once_with(
             service="s3",
             command="put_object",
@@ -96,12 +85,10 @@ class TestTemplate(object):
             expected_template_key
         )
 
-    def test_create_bucket_with_bucket_that_exists(self):
+    def test_bucket_exists_with_bucket_that_exists(self):
         # connection_manager.call doesn't raise an exception, mimicing the
         # behaviour when head_bucket successfully executes.
-        self.template._create_bucket(
-            "region", "bucket_name", self.connection_manager
-        )
+        self.template._bucket_exists("bucket_name", self.connection_manager)
 
     def test_create_bucket_with_unreadable_bucket(self):
         self.connection_manager.call.side_effect = ClientError(
@@ -120,7 +107,7 @@ class TestTemplate(object):
             assert e.value.response["Error"]["Code"] == 500
             assert e.value.response["Error"]["Message"] == "Bucket Unreadable"
 
-    def test_create_bucket_with_non_existent_bucket(self):
+    def test_bucket_exists_with_non_existent_bucket(self):
         # connection_manager.call is called twice, and should throw the
         # Not Found ClientError only for the first call.
         self.connection_manager.call.side_effect = [
@@ -136,38 +123,16 @@ class TestTemplate(object):
             None
         ]
 
-        self.template._create_bucket(
-            self.region,
+        existance = self.template._bucket_exists(
             self.bucket_name,
             self.connection_manager
         )
 
-        self.connection_manager.call.assert_any_call(
-            service="s3",
-            command="create_bucket",
-            kwargs={
-                "Bucket": self.bucket_name,
-                "CreateBucketConfiguration": {
-                    "LocationConstraint": self.region
-                },
-            }
-        )
+        assert existance is False
 
-    def test_create_bucket_with_non_existent_us_east_1_bucket(self):
+    def test_create_bucket_in_us_east_1(self):
         # connection_manager.call is called twice, and should throw the
         # Not Found ClientError only for the first call.
-        self.connection_manager.call.side_effect = [
-            ClientError(
-                {
-                    "Error": {
-                        "Code": 404,
-                        "Message": "Not Found"
-                    }
-                },
-                sentinel.operation
-            ),
-            None
-        ]
 
         self.template._create_bucket(
             "us-east-1",
@@ -181,73 +146,60 @@ class TestTemplate(object):
             kwargs={"Bucket": self.bucket_name}
         )
 
-    def test_get_cfn_with_json_template(self):
+    def test_body_with_json_template(self):
         self.template.name = "vpc"
         self.template.path = os.path.join(
             os.getcwd(),
             "tests/fixtures/templates/vpc.json"
         )
-        output = self.template._get_cfn()
+        output = self.template.body
         output_dict = json.loads(output)
         with open("tests/fixtures/templates/compiled_vpc.json", "r") as f:
             expected_output_dict = json.loads(f.read())
         assert output_dict == expected_output_dict
 
-    def test_get_cfn_with_yaml_template(self):
+    def test_body_with_yaml_template(self):
         self.template.name = "vpc"
         self.template.path = os.path.join(
             os.getcwd(),
             "tests/fixtures/templates/vpc.yaml"
         )
-        output = self.template._get_cfn()
+        output = self.template.body
         output_dict = yaml.load(output)
         with open("tests/fixtures/templates/compiled_vpc.json", "r") as f:
             expected_output_dict = json.loads(f.read())
         assert output_dict == expected_output_dict
 
-    def test_get_cfn_with_missing_file(self):
+    def test_body_with_missing_file(self):
         self.template.path = "incorrect/template/path.py"
         with pytest.raises(IOError):
-            self.template._get_cfn()
+            self.template.body
 
-    def test_get_cfn_compiles_troposphere(self):
+    def test_body_with_troposphere(self):
         self.template.sceptre_user_data = None
         self.template.name = "vpc"
         self.template.path = os.path.join(
             os.getcwd(),
             "tests/fixtures/templates/vpc.py"
         )
-        output = self.template._get_cfn()
-        try:
-            json.loads(output)
-        except:
-            assert False
-
-    def test_get_cfn_with_troposphere(self):
-        self.template.sceptre_user_data = None
-        self.template.name = "vpc"
-        self.template.path = os.path.join(
-            os.getcwd(),
-            "tests/fixtures/templates/vpc.py"
-        )
-        actual_output = json.loads(self.template._get_cfn())
+        actual_output = json.loads(self.template.body)
         with open("tests/fixtures/templates/compiled_vpc.json", "r") as f:
             expected_output = json.loads(f.read())
         assert actual_output == expected_output
 
-    def test_get_cfn_with_troposphere_with_sgt(self):
+    def test_body_with_troposphere_with_sgt(self):
         self.template.sceptre_user_data = None
         self.template.name = "vpc_sgt"
         self.template.path = os.path.join(
             os.getcwd(),
             "tests/fixtures/templates/vpc_sgt.py"
         )
-        actual_output = json.loads(self.template._get_cfn())
+        actual_output = json.loads(self.template.body)
         with open("tests/fixtures/templates/compiled_vpc.json", "r") as f:
             expected_output = json.loads(f.read())
         assert actual_output == expected_output
 
-    def test_get_cfn_injects_sceptre_user_data(self):
+    def test_body_injects_sceptre_user_data(self):
         self.template.sceptre_user_data = {
             "cidr_block": "10.0.0.0/16"
         }
@@ -257,12 +209,12 @@ class TestTemplate(object):
             "tests/fixtures/templates/vpc_sud.py"
         )
 
-        actual_output = json.loads(self.template._get_cfn())
+        actual_output = json.loads(self.template.body)
         with open("tests/fixtures/templates/compiled_vpc_sud.json", "r") as f:
             expected_output = json.loads(f.read())
         assert actual_output == expected_output
 
-    def test_get_cfn_injects_sceptre_user_data_incorrect_function(self):
+    def test_body_injects_sceptre_user_data_incorrect_function(self):
         self.template.sceptre_user_data = {
             "cidr_block": "10.0.0.0/16"
         }
@@ -272,9 +224,9 @@ class TestTemplate(object):
             "tests/fixtures/templates/vpc_sud_incorrect_function.py"
         )
         with pytest.raises(TemplateSceptreHandlerError):
-            self.template._get_cfn()
+            self.template.body
 
-    def test_get_cfn_injects_sceptre_user_data_incorrect_handler(self):
+    def test_body_injects_sceptre_user_data_incorrect_handler(self):
         self.template.sceptre_user_data = {
             "cidr_block": "10.0.0.0/16"
         }
@@ -284,11 +236,11 @@ class TestTemplate(object):
             "tests/fixtures/templates/vpc_sud_incorrect_handler.py"
         )
         with pytest.raises(TypeError):
-            self.template._get_cfn()
+            self.template.body
 
-    def test_get_cfn_with_incorrect_filetype(self):
+    def test_body_with_incorrect_filetype(self):
         self.template.path = (
             "path/to/something.ext"
         )
         with pytest.raises(UnsupportedTemplateFileTypeError):
-            self.template._get_cfn()
+            self.template.body